seq: use guint8 for the numerical ID of client

+++ src/seq/query.c
@@ -259,7 +259,7 @@ void alsaseq_get_client_id_list(guint **entries, gsize *entry_count,
  * with SNDRV_SEQ_IOCTL_GET_CLIENT_INFO command for ALSA sequencer character
  * device.
  */
-void alsaseq_get_client_info(guint client_id, ALSASeqClientInfo **client_info,
+void alsaseq_get_client_info(guint8 client_id, ALSASeqClientInfo **client_info,
                              GError **error)
 {
     char *devnode;
@@ -309,7 +309,7 @@ void alsaseq_get_client_info(guint client_id, ALSASeqClientInfo **client_info,
  * with SNDRV_SEQ_IOCTL_GET_CLIENT_INFO and SNDRV_SEQ_IOCTL_QUERY_NEXT_PORT
  * commands for ALSA sequencer character device.
  */
-void alsaseq_get_port_id_list(guint client_id, guint **entries,
+void alsaseq_get_port_id_list(guint8 client_id, guint **entries,
                               gsize *entry_count, GError **error)
 {
     char *devnode;
@@ -385,7 +385,7 @@ void alsaseq_get_port_id_list(guint client_id, guint **entries,
  * with SNDRV_SEQ_IOCTL_GET_PORT_INFO command for ALSA sequencer character
  * device.
  */
-void alsaseq_get_port_info(guint client_id, guint port_id,
+void alsaseq_get_port_info(guint8 client_id, guint port_id,
                            ALSASeqPortInfo **port_info, GError **error)
 {
     char *devnode;
@@ -433,7 +433,7 @@ void alsaseq_get_port_info(guint client_id, guint port_id,
  * with SNDRV_SEQ_IOCTL_GET_CLIENT_POOL command for ALSA sequencer character
  * device.
  */
-void alsaseq_get_client_pool(gint client_id, ALSASeqClientPool **client_pool,
+void alsaseq_get_client_pool(guint8 client_id, ALSASeqClientPool **client_pool,
                              GError **error)
 {
     char *devnode;
